property_tracker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
jinja2==3.1.2
aiofiles==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
boto3==1.34.0
psycopg2-binary==2.9.9